Fix test_stack_context for Tornado 3
From Ben Darnell: "This error has to do with the relationship between the stack_context and the actual stack.  If a coroutine is suspended during the "with" statement that establishes a stack context, then the exit half of the "with" will not be run until after the coroutine is resumed.  You can't do anything asynchronous inside the with block, although you could schedule pool.get_socket on the IOLoop with add_callback."