MOTOR-38, fix tests for narrowed localhost exception.
See SERVER 12621, nearly the only thing allowed on MongoDB 2.7.1+,
for a server started with auth but with no users, is to create an admin
user. Motor's tests now detect this case and immediately create an
admin user. The test MotorClient remains logged in as the admin user
throughout the suite.