Optionally test with pytest.
Useful if we want to output JUnit XML for Evergreen to read. But since
we don't use our custom test runner defined in setup.py anymore, we must
rename TestEnvironment to Environment so it isn't interpreted as a test,
and declare all our test dependencies in tox.ini, since we can't rely on
setup.py to install its tests_require dependencies.