MOTOR-209 Wrap ClientSession in Motor class
In Motor 1.2 we supported logical sessions using the ClientSession class
from PyMongo - there was no need to wrap it because it does no I/O and
has no async methods of its own.

Now, closing a session that is in a transaction sends a command that
commits or aborts the transaction, and the application must await the
outcome. Therefore, I wrap ClientSession in a new MotorClientSession
class and apps must use it in "async with" instead of "with".

This incompatible change motivates Motor 2.0.