Proposed changes to motor examples
Hi,

For a newbie to motor and asyncio the code proposed here would not work especially for python3.9 onwards. asyncio automatically closes the event loop at the end of asyncio.run() method. Subsequently the aiohttp is not able to open new event loop[issue is faced on all platforms: windows, linux, mac_os]. Proposed are the changes works and ensures that the event loop is consistent among aiohttp and motor's event loop